+++ src/leiningen/new/auth_base.clj
@@ -7,8 +7,8 @@
      (-> options
          (assoc :auth true)
          (append-options :dependencies [['buddy/buddy-auth "2.2.0"]
-                                        ['buddy/buddy-sign "3.1.0"]
-                                        ['buddy/buddy-core "1.6.0"]
+                                        ['buddy/buddy-sign "3.2.0"]
+                                        ['buddy/buddy-core "1.8.0"]
                                         ['buddy/buddy-hashers "1.6.0"]])
          (append-formatted :auth-middleware-required
                            [['buddy.auth.middleware :refer ['wrap-authentication 'wrap-authorization]]