+++ src/leiningen/new/db.clj
@@ -11,9 +11,9 @@
     (some #{"+sqlite"} features) :sqlite))
 
 (defn db-dependencies [options]
-  (into [['luminus-migrations "0.6.7"]
+  (into [['luminus-migrations "0.6.9"]
          ['conman "0.9.0"]]
-        ({:postgres [['org.postgresql/postgresql "42.2.14"]]
+        ({:postgres [['org.postgresql/postgresql "42.2.18"]]
           :mysql    [['mysql/mysql-connector-java "8.0.18"]
                      ['com.google.protobuf/protobuf-java "3.8.0"]]
           :h2       [['com.h2database/h2 "1.4.200"]]