+++ src/leiningen/new/hoplon.clj
@@ -9,7 +9,7 @@
   [['hoplon "7.2.0"]
    ['cljsjs/jquery "3.2.1-0"]
    ['clj-commons/secretary "1.2.4"]
-   ['cljs-ajax "0.8.0"]])
+   ['cljs-ajax "0.8.1"]])
 
 (defn hoplon-features [[assets options :as state]]
   (if (some #{"+hoplon"} (:features options))