+++ src/leiningen/new/reitit.clj
@@ -9,8 +9,8 @@
   (if (some #{"+reitit"} (:features options))
     [(into (remove-conflicting-assets assets "home.clj" "docs.md") reitit-assets)
      (-> options
-         (append-options :dependencies [['metosin/reitit "0.5.5"]
+         (append-options :dependencies [['metosin/reitit "0.5.9"]
                                         ['metosin/jsonista "0.2.7"]
-                                        ['com.fasterxml.jackson.core/jackson-core "2.11.2"]])
+                                        ['com.fasterxml.jackson.core/jackson-core "2.11.3"]])
          (assoc :reitit true))]
     state))